Move types from navbar to topbar
