+++ packages/app/src/features/topbar/logic/useTopbar.ts
@@ -3,7 +3,7 @@ import { useSandboxState } from '@/domain/sandbox/useSandboxState'
 import { useOpenDialog } from '@/domain/state/dialogs'
 import { sandboxDialogConfig } from '@/features/dialogs/sandbox/SandboxDialog'
 import { selectNetworkDialogConfig } from '@/features/dialogs/select-network/SelectNetworkDialog'
-import { SupportedChain } from '@/features/navbar/types'
+import { SupportedChain } from '@/features/topbar/types'
 import { getBuildInfo } from '@/ui/utils/getBuildInfo'
 import { raise } from '@/utils/assert'
 import { CheckedAddress } from '@marsfoundation/common-universal'
